Temel CBS kabiliyetleri eklenmiştir. OpenGL Tabanlı painter eklenmiştir. ImGUI eklenmiştir. Ornek google harita paftalari eklenmiştir.

+++ code/src/infra/application_base/mouse_move_event_listener.h
@@ -13,7 +13,7 @@
 
 class MouseMoveEventListener {
 public:
-	virtual void MouseMoveEvent(double x, double y) = 0;
+	virtual void MouseMoveEvent(uint16_t x, uint16_t  y) = 0;
 };
 
 /**
